Fix compiler errors and warnings:

+++ src/ds3.h
@@ -243,7 +243,7 @@ LIBRARY_API size_t ds3_write_to_file(void* buffer, size_t size, size_t nmemb, vo
 LIBRARY_API size_t ds3_read_from_file(void* buffer, size_t size, size_t nmemb, void* user_data);
 
 LIBRARY_API ds3_bulk_object_list* ds3_convert_file_list(const char** file_list, uint64_t num_files);
-LIBRARY_API ds3_bulk_object_list* ds3_convert_object_list(const ds3_object* objects, size_t num_objects);
+LIBRARY_API ds3_bulk_object_list* ds3_convert_object_list(const ds3_object* objects, uint64_t num_objects);
 LIBRARY_API ds3_bulk_object_list* ds3_init_bulk_object_list(uint64_t num_files);
 LIBRARY_API void ds3_free_bulk_object_list(ds3_bulk_object_list* object_list);
